pyproject.toml
[project]
name = "skillforge"
version = "0.1.0"
description = "Generate agent skills from documentation using Firecrawl"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.77.0",
    "click>=8.3.1",
    "firecrawl-py>=4.14.0",
    "rich>=14.3.1",
]

[project.scripts]
skillforge = "skillforge.cli:main"

[tool.uv]
package = true

[tool.setuptools]
packages = ["skillforge"]
